+++ packages/music/CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 > `Security` - in case of vulnerabilities.
 
-## [1.3] - 2021-10-29
+## [1.3.1] - 2021-10-29
 
 ### Added