+++ packages/utilities/CHANGELOG.md
@@ -19,6 +19,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 > `Security` - in case of vulnerabilities.
 
+## [3.0.0] - 2021-10-29
+
+### Added
+
+- Early exit of pagination (before timeout)
+- Throw error if pagination already sent
+
+### Changed
+
+- paginations options
+
 ## [2.1.1] - 2021-10-28
 
 ### Fixed